organized sources in detail/meta dir
