+++ include/fkYAML/detail/conversions/from_node.hpp
@@ -17,6 +17,8 @@
 #include <utility>
 
 #include <fkYAML/detail/macros/version_macros.hpp>
+#include <fkYAML/detail/meta/node_traits.hpp>
+#include <fkYAML/detail/meta/stl_supplement.hpp>
 #include <fkYAML/detail/meta/type_traits.hpp>
 #include <fkYAML/detail/types/node_t.hpp>
 #include <fkYAML/exception.hpp>
