+++ include/fkYAML/detail/conversions/to_node.hpp
@@ -17,7 +17,9 @@
 
 #include <fkYAML/detail/macros/version_macros.hpp>
 #include <fkYAML/detail/types/node_t.hpp>
+#include <fkYAML/detail/meta/node_traits.hpp>
 #include <fkYAML/detail/meta/type_traits.hpp>
+#include <fkYAML/detail/meta/stl_supplement.hpp>
 
 /**
  * @namespace fkyaml
