+++ include/fkYAML/detail/iterator.hpp
@@ -17,7 +17,7 @@
 #include <iterator>
 
 #include <fkYAML/detail/macros/version_macros.hpp>
-#include <fkYAML/detail/meta/type_traits.hpp>
+#include <fkYAML/detail/meta/node_traits.hpp>
 #include <fkYAML/exception.hpp>
 
 /**
